% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_temperature}
\alias{delta_temperature}
\title{Delta Temperature}
\format{
a 3 dimensional array [12 months, 22 years, 2 deltas]
}
\source{
Data Wrangling and Additional Temperature Modeling: Sadie Gill  \email{sgill@flowwest.com}
}
\usage{
delta_temperature
}
\description{
The 1979-2000 rearing temperature in the North and South Delta in °C
}
\details{
Water temperature in the Deltas were modeled as a function of air temperature.

North Delta was modeled using a sample of recent measured water temperature at
CDEC station \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=EMM}{EMM}
and NOAA CDO air temperature data from \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00040232/detail}{Antioch}.

The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

South Delta was modeled using a sample of recent measured water temperature at
CDEC station \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=MTB}{MTB}
and NOAA CDO air temperature data \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023237/detail}{Stockton Airport}.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.

Air Temperature Data Source:
 \itemize{
   \item NOAA Climate Data Online \href{https://www.ncdc.noaa.gov/cdo-web/}{(CDO)} accessed using the \href{https://github.com/ropensci/rnoaa}{\code{rnoaa}} R package developed by \href{https://ropensci.org/}{rOpenSci}.
   }

 Stream Water Temperature Data Sources:
 \itemize{
   \item DWR California Data Exchange Center \href{http://cdec.water.ca.gov/index.html}{(CDEC)} accessed using the \href{https://github.com/flowwest/cdecretrieve}{\code{CDECRetrieve}} R packaged developed by \href{http://www.flowwest.com/}{FlowWest}.
 }
}
\keyword{datasets}
