% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_egg_temperature_effect}
\alias{fr_egg_temperature_effect}
\title{Temperature Effect on Egg Mortality}
\format{
A length 31 vector
}
\usage{
fr_egg_temperature_effect
}
\description{
The mean estimate of temperature effect used for egg to fry survival
}
\details{
Estimated by C. Hammersmark (CBEC ECOengineering Inc.).
Was calculated by taking the mean of dry and wet egg temp effects from previous model.
}
\keyword{datasets}
