% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migratory_temperature_proportion_over_20}
\alias{migratory_temperature_proportion_over_20}
\title{Migratory Corridor Water Temperature}
\format{
a 31 by 12 matrix [watershed by month]
}
\source{
\itemize{
  \item Method developed by James T. Peterson \email{jt.peterson@oregonstate.edu}
  \item Data wrangling by Sadie Gill \email{sgill@flowwest.com}
}
}
\usage{
migratory_temperature_proportion_over_20
}
\description{
The median proportion of days over 20°C per month
}
\details{
Watersheds were assigned membership to one of the following three locations to
represent the migratory corridor temperature conditions. The data was summarised
by calculating the median proportion of days over 20°C for each month during the available period of record.

Sacramento River, \href{http://cdec.water.ca.gov/dynamicapp/staMeta?station_id=WLK}{Sacramento River Below Wilkins Slough}:
Upper Sacramento River, Antelope Creek, Battle Creek, Bear Creek, Big Chico Creek Butte Creek, Clear Creek, Cottonwood Creek,
Cow Creek, Deer Creek, Elder Creek, Mill Creek, Paynes Creek, Stony Creek, Thomes Creek, Bear River, Feather River, Yuba River
American River

South Delta, \href{http://cdec.water.ca.gov/dynamicapp/staMeta?station_id=MOK}{Mokelumne River At San Joaquin River}:
Calaveras River, Cosumnes River, Mokelumne River

San Joaquin River, \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11303500}{San Joaquin River Near Vernalis}:
Merced River, Stanislaus River, Tuolumne River

No spawning in these regions (assigned value of zero):
Upper-mid Sacramento River, Sutter Bypass, Lower-mid Sacramento River, Yolo Bypass, Lower Sacramento River, San Joaquin River
}
\keyword{datasets}
