% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream_temperature}
\alias{stream_temperature}
\title{Monthly Mean Water Temperature}
\format{
A named list with two components, each containing a 3 dimensional array [31 watersheds, 12 months, 21 years]
}
\source{
\itemize{
  \item HEC-5Q Model Output and Double Regression Temperature Modeling: Michael Wright \email{mwright@usbr.gov}
  \item Data Wrangling and Additional Temperature Modeling: Sadie Gill  \email{sgill@flowwest.com}
}
}
\usage{
stream_temperature
}
\description{
A named list containing the 1980-2000 monthly mean water temperatures in °C.
Each named list element is a temperature datasets of Hec5q and other modeled temps corresponding to 2008-2009 and 2018-2019 Calsim II BiOp runs.

The named list components are:
\itemize{
    \item{biop_2008_2009: monthly stream temperature in degrees C for watersheds from the 2008-2009 Calsim II BiOp run.}
    \item{biop_itp_2018_2019: monthly stream temperature in degrees C for watersheds from the 2018-2019 Calsim II BiOp & ITP run.}
    \item{run_of_river: monthly stream temperature in degrees C for watersheds from the Run of River Calsim run.}
}
}
\details{
The following four methods were used to estimate the monthly mean water
temperature for the watersheds during 1979-1999:


\strong{HEC-5Q} \cr
Temperature assignments were determined by Mike Wright using the input files
for the three HEC-5Q runs, i.e. AR_5QCS.dat and AR_5Q-CL.OUT.


\strong{Empirical Data} \cr
Only the Lower Sacramento had sufficient measured water temperature during
the period of the R2R salmon life cycle model. The few missing values were
imputed using \href{https://www.rdocumentation.org/packages/forecast/versions/8.1/topics/na.interp}{\code{forecast::na.interp}}.

 \strong{Double Regression Modeling} \cr
 Data from many temperature gauges was collected from several R2R watersheds.
 In these watersheds, a double regression technique was employed to predict water
 temperature based on equilibrium temperature and the river mile of interest.
 Equilibrium temperature is the temperature a pool of water would approach if
 left in contact with the air for an infinite time, generally equivalent to the
 air temperature. The results were applied to both watersheds of origin and as
 surrogate data for ungauged watersheds.

 \href{https://cvpiatemperature-r-package.s3-us-west-2.amazonaws.com/Temperature_Regression_Model_for_CVPIA_Streams.pdf}{Modeling Details}

\strong{Additional Temperature Modeling} \cr
For streams without major dams, water temperature is highly correlated with air temperature.
For each watershed not included in the HEC-5Q model that had partial water temperature data,
a linear model was fitted to estimate water temperature as a function of air temperature.

Generally, the air temperature record spans both the period of the R2R salmon life cycle model
and the complete period of record of available water temperature data. In the
cases where there were missing air temperature values between 1979-1999, we imputed the missing values using
\href{https://www.rdocumentation.org/packages/forecast/versions/8.1/topics/na.interp}{\code{forecast::na.interp}}
in order to have a complete air temperature dataset for water temperature estimation.

 Air Temperature Data Source:
 \itemize{
   \item NOAA Climate Data Online \href{https://www.ncdc.noaa.gov/cdo-web/}{(CDO)} accessed using the \href{https://github.com/ropensci/rnoaa}{\code{rnoaa}} R package developed by \href{https://ropensci.org/}{rOpenSci}.
   }

 Stream Water Temperature Data Sources:
 \itemize{
   \item DWR California Data Exchange Center \href{http://cdec.water.ca.gov/index.html}{(CDEC)} accessed using the \href{https://github.com/flowwest/cdecretrieve}{\code{CDECRetrieve}} R packaged developed by \href{http://www.flowwest.com/}{FlowWest}.
   \item USGS National Water Information System \href{https://waterdata.usgs.gov/nwis}{(NWIS)} accessed using the \href{https://github.com/USGS-R/dataRetrieval}{\code{dataRetrieval}} R package developed by \href{https://www.usgs.gov/}{USGS}.
 }
}
\section{Watershed Modeling Details}{

\itemize{
   \item \strong{Upper Sacramento River}
     \itemize{
       \item{biop_2008_2009: HEC-5Q model output}
       \item{biop_itp_2018_2019: HEC-5Q model output}
       \item{run_of_river: Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{# https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=jlf}{JLF} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024216/detail}{USW00024216}}
   }

   \item \strong{Antelope Creek} Regression method was fitted to data from in-river temperature gauges
   \item \strong{Battle Creek}
     \itemize{
       \item{biop_2008_2009: HEC-5Q model output}
       \item{biop_itp_2018_2019: HEC-5Q model output}
       \item{run_of_river: Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{# https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=bnf}{BNF} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024257/detail}{USW00024257}}
   }
   \item \strong{Bear Creek} Regression method was fitted to data from in-river temperature gauges
   \item \strong{Big Chico Creek} Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{http://cdec4gov.water.ca.gov/dynamicapp/QueryF?s=BIC}{BIC} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00046685/detail}{USC00046685}
   \item \strong{Butte Creek} Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=BCD}{BCD} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00046685/detail}{USC00046685}
   \item \strong{Clear Creek} HEC-5Q model  output at IGO in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS
   \item \strong{Cottonwood Creek} HEC-5Q model output at COTTONWOOD CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS
   \item \strong{Cow Creek} HEC-5Q model  output at COW CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS
   \item \strong{Deer Creek} Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11383500}{11383500} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00041715/detail}{USC00041715}
   \item \strong{Elder Creek} Regression method was fitted to data from Antelope River temperature gauges
   \item \strong{Mill Creek} Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11381500}{11381500} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024216/detail}{USW00024216}
   \item \strong{Paynes Creek} Regression method was fitted to data from Antelope River temperature gauges
   \item \strong{Stony Creek} HEC-5Q model  output at STONY CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS
   \item \strong{Thomes Creek} HEC-5Q model output at THOMES CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS
   \item \strong{Upper-mid Sacramento River}
    \itemize{
      \item{biop_2008_2009: HEC-5Q model output}
       \item{biop_itp_2018_2019: HEC-5Q model output}
       \item{run_of_river: Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{# https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=rdb}{RDB} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024216/detail}{USW00024216}}
   }
   \item \strong{Sutter Bypass} Tisdale Rotary Screw Trap monthly mean water temperature 2011-2017
   \item \strong{Bear River} Regression method was fitted to data from Deer Creek temperature gauges
   \item \strong{Feather River} Regression method was fitted to data from Deer Creek temperature gauges
   \item \strong{Yuba River} Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11421000}{11421000} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024216/detail}{USW00024216}
   \item \strong{Lower-mid Sacramento River}
     \itemize{
       \item{biop_2008_2009: HEC-5Q model output}
       \item{biop_itp_2018_2019: HEC-5Q model output}
       \item{run_of_river: Estimated mean monthly water temperature from a mean monthly water temperature linear model fitted with water temperature from USGS Gage ID:\href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11391100}{11391100} with imputed missing values using \href{https://www.rdocumentation.org/packages/forecast/versions/8.1/topics/na.interp}{\code{forecast::na.interp}}}
     }
   \item \strong{Yolo Bypass} Knights Landing Rotary Screw Trap monthly mean water temperature 2004-2018
   \item \strong{American River} HEC-5Q model output at WILLIAM POND PARK in SSJB_AR_Reference_063015/AR/AR_CL_Temp.dss
   \item \strong{Lower Sacramento River} Measured mean monthly water temperature from  USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11447650}{11447650} with imputed missing values using \href{https://www.rdocumentation.org/packages/forecast/versions/8.1/topics/na.interp}{\code{forecast::na.interp}}
   \item \strong{Calaveras River} Regression method was fitted to data from Deer Creek temperature gauges
   \item \strong{Cosumnes River} Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11335000}{11335000} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023271/detail}{USW00023271}
   \item \strong{Mokelumne River} Estimated mean monthly water temperature from a linear model fitted with water temperature data provided by EBMUD measured near Victor, CA and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00045032/detail}{USC00045032}
   \item \strong{Merced River}
       \itemize{
           \item{biop_2008_2009: HEC-5Q model output}
           \item{biop_itp_2018_2019: HEC-5Q model for San Joaquin is no longer supported. Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage Id: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=CRS}{CRS} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00045532/detail}{USC00045532}}
           \item{run_of_river: used same water to air temperature regression data as biop_itp_2018_2019}
        }
   \item \strong{Stanislaus River}
        \itemize{
           \item{biop_2008_2009: HEC-5Q model output at BLW MCHENRY BR in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS}
           \item{biop_itp_2018_2019: HEC-5Q model for San Joaquin is no longer supported. Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage Id: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=RPN}{RPN} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023258/detail}{USW00023258}}
           \item{run_of_river: used same water to air temperature regression data as biop_itp_2018_2019}
        }
   \item \strong{Tuolumne River}
         \itemize{
           \item{biop_2008_2009: HEC-5Q model output at GEER ROAD BR in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS}
           \item{biop_itp_2018_2019: HEC-5Q model for San Joaquin is no longer supported. Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage Id: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=MOD}{MOD} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023258/detail}{USW00023258}}
           \item{run_of_river: used same water to air temperature regression data as biop_itp_2018_2019}
        }
   \item \strong{San Joaquin River}
         \itemize{
           \item{biop_2008_2009: HEC-5Q model output at ABV TUOL in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS}
           \item{biop_itp_2018_2019: HEC-5Q model for San Joaquin is no longer supported. Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage Id: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=CRS}{CRS} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00045532/detail}{USC00045532}}
           \item{run_of_river: used same water to air temperature regression data as biop_itp_2018_2019}
        }
}
}

\keyword{datasets}
