% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vernalis_temperature}
\alias{vernalis_temperature}
\title{Vernalis Water Temperature}
\format{
A matrix 12x21 [12 months by 21 years]
}
\usage{
vernalis_temperature
}
\description{
Monthly water temperature (°C) at Vernalis from 1980-2000
}
\details{
Temperature at Vernalis was developed as a function of air
temperature. The most recent water temperature at San Joation River (SJR)
\href{https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=SJR}{CDEC: RRI} was used for water temperature. Air temperature was obtained using
NOAA CDO at \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023237/detail}{Stockton Metropolitan Airport}.
}
\keyword{datasets}
