% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wr_egg_temperature_effect}
\alias{wr_egg_temperature_effect}
\title{Temperature Effect on Egg Mortality}
\format{
A length 31 vector
}
\usage{
wr_egg_temperature_effect
}
\description{
The mean estimate of temperature effect used for egg to fry survival
}
\details{
Was calculated through model calibration.
}
\keyword{datasets}
