% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degree_days}
\alias{degree_days}
\title{Degree Days}
\format{
a 3 dimensional array [31 watersheds, 12 months, 22 years]
}
\source{
\itemize{
  \item HEC-5Q Model Output and Double Regression Temperature Modeling: Michael Wright \email{mwright@usbr.gov}
  \item Data Wrangling and Additional Temperature Modeling: Sadie Gill  \email{sgill@flowwest.com}
}
}
\usage{
degree_days
}
\description{
The monthly accumulated degree days °C
}
\details{
For watersheds with HEC-5Q modeled results, the calculation for degree days
is the sum of the daily mean values for each month.

For the other regions, the calculation for degree days is the monthly mean
temperature multiplied by the number of days in the month.

For more details about the temperature modeling see \code{\link{stream_temperature}}
}
\keyword{datasets}
