% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egg_temperature_effect}
\alias{egg_temperature_effect}
\title{Temperature Effect on Egg Mortality}
\format{
A dataframe with 4 columns (\code{watershed}, \code{fall_run}, \code{spring_run}, \code{winter_run}), and 31 rows (a row for each watershed).
}
\usage{
egg_temperature_effect
}
\description{
The mean estimate of temperature effect used for egg to fry survival
}
\details{
Fall and Spring Run values are the same and were estimated by C. Hammersmark (CBEC ECOengineering Inc.).
These Fall and Spring Run values were calculated by taking the mean of dry and wet egg temp effects from previous model.
Winter run values were calculated through model calibration.
}
\keyword{datasets}
